#1818 | reorganize top-sheet files
