+++ client/elements/addons/sc-top-sheet-info.js
@@ -1,9 +1,9 @@
 import {html, css, LitElement} from 'lit-element';
-import '../../suttaplex/card/sc-suttaplex.js';
+import '../suttaplex/card/sc-suttaplex.js';
 import {unsafeHTML} from "lit-html/directives/unsafe-html";
-import SCSuttaTopSheetCommon from './sc-top-sheet-common';
+import SCTopSheetCommon from './sc-top-sheet-common';
 
-class SCSuttaInfo extends SCSuttaTopSheetCommon {
+class SCTopSheetInfo extends SCTopSheetCommon {
 
     constructor() {
         super();
@@ -78,4 +78,4 @@ class SCSuttaInfo extends SCSuttaTopSheetCommon {
     }
 }
 
-customElements.define('sc-sutta-info', SCSuttaInfo);
+customElements.define('sc-top-sheet-info', SCTopSheetInfo);
