+++ client/elements/addons/sc-top-sheet-parallels.js
@@ -1,9 +1,9 @@
 import {html, css} from 'lit-element';
-import '../../suttaplex/card/sc-suttaplex.js';
-import { suttaplexStyles } from '../../styles/sc-suttaplex-styles.js';
-import SCSuttaTopSheetCommon from "./sc-top-sheet-common";
+import '../suttaplex/card/sc-suttaplex.js';
+import { suttaplexStyles } from '../styles/sc-suttaplex-styles.js';
+import SCTopSheetCommon from "./sc-top-sheet-common";
 
-class SCSuttaParallels extends SCSuttaTopSheetCommon {
+class SCTopSheetParallels extends SCTopSheetCommon {
 
   static get styles() {
     return css`
@@ -63,4 +63,4 @@ class SCSuttaParallels extends SCSuttaTopSheetCommon {
   }
 }
 
-customElements.define('sc-sutta-parallels', SCSuttaParallels);
+customElements.define('sc-top-sheet-parallels', SCTopSheetParallels);
